+++ src/components/indicatorCard/IndicatorCard.module.scss
