+++ src/components/indicatorCard/IndicatorCard.tsx
